1、移动签名方法到单独的类；2、整理网络请求方法和注释；
