+++ app/src/main/java/com/byl/mvvm/api/response/BaseResult.kt
@@ -4,4 +4,8 @@ open class BaseResult<T> {
     val errorMsg: String? = null
     val errorCode: Int = 0
     val data: T? = null
+
+    fun isSuccess(): Boolean {
+        return errorCode == 0
+    }
 }
